Linux Foundation bug 1505: Findif parsing is all screwed up unless all the parameters are passed, not defaulted.
I changed findif to use the OCF parameters that are passed to
both IPaddr and IPaddr2, which got rid of a bunch of ugly code
at the same time as fixing this bug.  Sounds like a win to me.

In the process I had to make IPaddr and IPaddr2 somewhat more consistent
than they had been - but not in an incompatible way, and this is also
a good thing.

Mercurial revision: 187c17c0769e7c659d519944b0821ea4939cd515