[RPM] Exclude ldirectord from the heartbeat package
The ldirectord heartbeat resource should be part of the
ldirectord package and not part of the heartbeat package.

This config change, borrowed from the SuSE .spec file,
ensures that ldirectord is not included in the hearbteat package.

It is already listed for inclusion in the ldirectord package.

Signed-off-by: Simon Horman <horms@verge.net.au>

Mercurial revision: 2de2745b8bf8eec9403522ce33b80f04e1d9a5e9