LRM (LF 1684): a meta-data operation fix introducing short sleep on zero reads
A meta-data operation used to sleep for one second in case output
from the OCF RA was not ready. This would make lrmd do nothing
for that period of time and in an unfortunate timing on signal
delivery immediately before, the signal handling would be
delayed. Thus the annoying message:

G_SIG_dispatch: Dispatch function for SIGCHLD was delayed 1000 ms (> 100 ms)

This patch should improve the situation significantly, but the
proper way would be to read the meta-data output in an
asynchronous manner, either in a child process or by using a
proper glib facility.

Thanks to Alan Robertson for spotting the problem.

Mercurial revision: 0ded50597e97b947c03e0090367f4f608e49f820