heartbeat: Increase MAXMSGHIST; reduce debugging output.
In a 7 node cluster (observed under CTS), the DC will experience
significant message bursts when the TE rolls out the actions. We've got
several clones, resulting in ~269 actions (handed to the message layer
in one go); our MSGHIST only allows for 200, and we logged copious
amounts of debugging output at 3/4 of that for _each_ message from there
on.

This patch increases MAXMSGHIST, which might have impact on the
flowcontrol code (which I'll be testing).

It makes the logging in hist_display() (which occured at LOG_DEBUG
anyway) dependent on ANYDEBUG; that ought to be perfectly safe, and that
definitely flooded the logging buffers. (If CTS is run with debug 1,
it'll still kill CTS; is this logging needed at all still?)

Mercurial revision: 69f0395c2eadf224c520a693b5eb115b01ab61bf