High: Tools: crm: improve processing of group edit and constraints
The set of changes to be applied to the CIB is at times
impossible if there are constraints which reference the affected
elements. This patch also makes possible moving resources between
groups and adding resources to a group. The commit processing is
made more clear with addition of "moved" and "recreate" options
to CIB element objects.

Mercurial revision: b8a6b31c5bdc2de3dbf5e60eba3d76eead25bf12