High: legacy: Set to the minimum scheduling priority when using SCHED_RR policy (bnc#779259)
Corosync-1.4.x defaults to use SCHED_RR scheduling policy with the
maximum priority. In the plugin mode, the pacemaker's daemons inherit
this, and they have the higher priority than sbd. If there are a lot of
actions in the cluster, crmd and cib will hog the CPUs, and it's
possible that sbd will not be scheduled for quite some time, which will
cause the watchdog timer expires.

The patch sets to the minimum scheduling priority for pacemaker's daemons
when using SCHED_RR policy.

Backported from: ClusterLabs/pacemaker@e79623059e07afec899a87d70376ce46c40c7e65